pyproject.toml
[tool.poetry]
name = "goalwise"
version = "0.1.0"
description = ""
authors = ["Clive"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
celery = "^5.3.4"
redis = "^5.0.1"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
